+++ Amperfy/Screens/Basics/Toolbar/NowPlayingBarItem.swift
@@ -11,167 +11,7 @@ import UIKit
 import AmperfyKit
 import MarqueeLabel
 
-protocol Refreshable {
-    func reload()
-}
-
-class ControlBarButton: UIBarButtonItem, MusicPlayable, Refreshable {
-    var player: PlayerFacade?
-
-    var inUIButton: UIButton? {
-        self.customView as? UIButton
-    }
-
-    func updateImage(image: UIImage) {
-        self.inUIButton?.configuration?.image = image.styleForNavigationBar()
-        //self.inUIButton?.setImage(image.styleForNavigationBar(), for: .normal)
-    }
-
-    fileprivate func createInUIButton(config: UIButton.Configuration, size: CGSize) -> UIButton? {
-        let button = UIButton(configuration: config)
-        button.imageView?.contentMode = .scaleAspectFit
-
-        // influence the highlighted area
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.widthAnchor.constraint(equalToConstant: size.width).isActive = true
-        button.heightAnchor.constraint(equalToConstant: size.height).isActive = true
-
-        return button
-    }
-
-    init(player: PlayerFacade) {
-        self.player = player
-        super.init()
-
-        var config = UIButton.Configuration.gray()
-        config.macIdiomStyle = .borderless
-        let button = createInUIButton(config: config, size: CGSize(width: 32, height: 22))
-        button?.addTarget(self, action: #selector(self.clicked(_:)), for: .touchUpInside)
-
-        self.customView = button
-
-        self.player?.addNotifier(notifier: self)
-
-        // Recreate the system button background highlight
-        self.installHoverGestureRecognizer()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func installHoverGestureRecognizer() {
-        let recognizer = UIHoverGestureRecognizer(target: self, action: #selector(self.hoverButton(_:)))
-        self.inUIButton?.addGestureRecognizer(recognizer)
-    }
-
-    @objc private func hoverButton(_ recognizer: UIHoverGestureRecognizer) {
-        switch recognizer.state {
-        case .began:
-            self.inUIButton?.backgroundColor = .systemGray2.withAlphaComponent(0.2)
-            self.inUIButton?.layer.cornerRadius = 5
-        case .ended, .cancelled, .failed:
-            self.inUIButton?.backgroundColor = .clear
-        default:
-            break
-        }
-    }
-
-    @objc func clicked(_ sender: UIButton) {
-
-    }
-
-    func reload() {}
-
-    func didStartPlaying() {}
-    func didPause() {}
-    func didStopPlaying() {}
-    func didStartPlayingFromBeginning() { }
-    func didElapsedTimeChange() {}
-    func didPlaylistChange() {}
-    func didArtworkChange() {}
-    func didShuffleChange() {}
-    func didRepeatChange() {}
-    func didPlaybackRateChange() {}
-}
-
-
-class PlayBarButton: ControlBarButton {
-    override var title: String? {
-        get { return "Play / Pause" }
-        set { }
-    }
-
-    override fileprivate func createInUIButton(config: UIButton.Configuration, size: CGSize) -> UIButton? {
-        // We need an initial image, otherwise catalyst calculates the wrong button size
-        var configuration = config
-        configuration.image = .play.styleForNavigationBar()
-        return super.createInUIButton(config: configuration, size: size)
-    }
-
-    override func clicked(_ sender: UIButton) {
-        self.player?.togglePlayPause()
-    }
-
-    override func didStartPlaying() {
-        self.updateImage(image: .pause)
-    }
-
-    override func didPause() {
-        self.updateImage(image: .play)
-    }
-
-    override func didStopPlaying() {
-        self.updateImage(image: .play)
-    }
-
-    override func reload() {
-        if player?.isPlaying ?? false {
-            self.updateImage(image: .pause)
-        } else {
-            self.updateImage(image: .play)
-        }
-    }
-}
-
-class NextBarButton: ControlBarButton {
-    override var title: String? {
-        get { return "Next" }
-        set { }
-    }
-
-    override fileprivate func createInUIButton(config: UIButton.Configuration, size: CGSize) -> UIButton? {
-        var configuration = config
-        configuration.image = .forwardFill.styleForNavigationBar(pointSize: 18)
-        // Increase the highlighted area
-        var newSize = size
-        newSize.width = 38
-        return super.createInUIButton(config: configuration, size: newSize)
-    }
-
-    override func clicked(_ sender: UIButton) {
-        self.player?.playNext()
-    }
-}
-
-class PreviousBarButton: ControlBarButton {
-    override var title: String? {
-        get { return "Previous" }
-        set { }
-    }
-
-    override fileprivate func createInUIButton(config: UIButton.Configuration, size: CGSize) -> UIButton? {
-        var configuration = config
-        configuration.image = .backwardFill.styleForNavigationBar(pointSize: 18)
-        var newSize = size
-        newSize.width = 38
-        return super.createInUIButton(config: configuration, size: newSize)
-    }
-
-    override func clicked(_ sender: UIButton) {
-        self.player?.playPreviousOrReplay()
-    }
-}
+#if targetEnvironment(macCatalyst)
 
 fileprivate class NowPlayingSlider: UISlider {
     static var sliderHeight: CGFloat = 5.0
@@ -324,7 +164,6 @@ fileprivate class NowPlayingInfoView: UIView {
             self.detailContainer.trailingAnchor.constraint(equalTo: self.trailingAnchor),
             self.detailContainer.heightAnchor.constraint(equalTo: self.heightAnchor, constant: 1)
         ])
-
         player.addNotifier(notifier: self)
         self.reload()
     }
@@ -436,27 +275,4 @@ class NowPlayingBarItem: UIBarButtonItem {
     }
 }
 
-
-// Hack: built-in flexible space navigation bar item is not working, so we use this workaround
-class FlexibleSpaceBarItem: UIBarButtonItem {
-    init(minSpace: CGFloat = 0, maxSpace: CGFloat = 1000) {
-        super.init()
-
-        let clearView = UIView(frame: .zero)
-        clearView.backgroundColor = .clear
-
-        clearView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            clearView.widthAnchor.constraint(lessThanOrEqualToConstant: maxSpace),
-            clearView.widthAnchor.constraint(greaterThanOrEqualToConstant: minSpace),
-            // This allows us to still grab and move the window when clicking the empty space
-            clearView.heightAnchor.constraint(equalToConstant: 0)
-        ])
-
-        self.customView = clearView
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
+#endif
